Move SQLExceptionMapper to db-protocol-core module
